+++ merger/column_alignment/hungarian.h
@@ -40,18 +40,18 @@ extern "C" {
 typedef struct {
   int num_rows;
   int num_cols;
-  int** cost;
+  double** cost;
   int** assignment;  
 } hungarian_problem_t;
 
 /** This method initialize the hungarian_problem structure and init 
  *  the  cost matrices (missing lines or columns are filled with 0).
  *  It returns the size of the quadratic(!) assignment matrix. **/
-int hungarian_init(hungarian_problem_t* p, 
-		   int** cost_matrix, 
-		   int rows, 
-		   int cols, 
-		   int mode);
+int hungarian_init(hungarian_problem_t* p,
+			double** cost_matrix,
+			int rows,
+			int cols,
+			int mode);
   
 /** Free the memory allocated by init. **/
 void hungarian_free(hungarian_problem_t* p);
